requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-PyMongo==2.3.0
pymongo==4.6.0
razorpay==1.4.1
python-dotenv==1.0.0
Werkzeug==2.3.7
gunicorn==21.2.0
# jwt==1.3.1
PyJWT==1.7.1
numpy==1.24.3
pillow
pytz
requests==2.31.0
urllib3==2.0.7